Extract Race enum
