+++ shared/Race.ts
@@ -1,0 +1,17 @@
+/**
+ * Reference list for all races.
+ *
+ * When referring to a race, it's preferred
+ * to use one of these values directly.
+ */
+export enum Race {
+  Dragonborn = `dragonborn`,
+  Dwarf = `dwarf`,
+  Elf = `elf`,
+  Gnome = `gnome`,
+  HalfElf = `half-elf`,
+  Halfling = `halfling`,
+  HalfOrc = `half-orc`,
+  Human = `human`,
+  Tiefling = `tiefling`,
+}
